Mv controls to map renderer
